+++ src/actions.rs
@@ -206,6 +206,7 @@ pub fn edit(
     notes: Option<&str>,
     folder_uuid: Option<&str>,
     history: &[crate::db::HistoryEntry],
+    key: Option<&str>,
 ) -> Result<(Option<String>, ())> {
     with_exchange_refresh_token(access_token, refresh_token, |access_token| {
         edit_once(
@@ -218,6 +219,7 @@ pub fn edit(
             notes,
             folder_uuid,
             history,
+            key,
         )
     })
 }
@@ -232,6 +234,7 @@ fn edit_once(
     notes: Option<&str>,
     folder_uuid: Option<&str>,
     history: &[crate::db::HistoryEntry],
+    key: Option<&str>,
 ) -> Result<()> {
     let (client, _) = api_client()?;
     client.edit(
@@ -244,6 +247,7 @@ fn edit_once(
         notes,
         folder_uuid,
         history,
+        key,
     )?;
     Ok(())
 }